Move NameListFormatter to its own class
